+++ cantera/trunk/Cantera/src/thermo/SimpleThermo.h
@@ -147,7 +147,7 @@ namespace Cantera {
      *
      * @see ConstCpPoly
      */
-    virtual void install(std::string name, size_t index, size_t type, const doublereal* c,
+    virtual void install(std::string name, size_t index, int type, const doublereal* c,
 			 doublereal minTemp, doublereal maxTemp, doublereal refPressure) {
 
       m_logt0.push_back(log(c[0]));
